nav bar fixes
nav bar now running header pic @ 28% width; nav links w/ 5% top/bottom
margin to center.

Will this have to be adjusted depending on how long the content on the
page is? e.g 5% is different depending on how long each snippet is?

and nav links are in the right order